opp_env,omnetpp: revise the way --local is implemented wrt download_commands and patch_commands
Removed "download_commands_local" and "patch_commands_local", and
instead the normal "download_commands" and "patch_commands" scripts
need to be written so that they are aware of whether --local is specified
or not. Thay can decide by examining the $LOCAL_OPERATION environmment
variable.

This fixes the bug that omnetpp .x versions could not be installed
using "--options=from-source-archive" and "--options=from-git"
because of failed patching