omnetpp_versions: resolving confusion about the 3.3p1 and 3.3p1-gcc73 releases
The original omnetpp-3.3 was released in 2006. Shortly after release, some
packaging problem was detected in the Windows build, and an update was
released as 3.3p1, only for Windows. Therefore, 3.3 and 3.3p1 are
essentially the same codebase.

In 2019, demand has arisen for a version that compiled with newer version.
The code was made to work with gcc-7.3 on Ubuntu 18.04, and the result
was released as 3.3p1-gcc73.

This release, in sync with similar changes in the omnetpp releases page
on github does the following:
- adds the original 3.3 aka 3.3p1 as 3.3, replacing the 3.3p1 label
- adds the 3.3p1-gcc73 release (from 2019) as 3.3.1 (kmm)